# api_gateway/docs/swagger.json
{
    "consumes": [
        "application/json",
        "application/x-www-form-urlencoded"
    ],
    "swagger": "2.0",
    "info": {
        "title": "API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/v1/",
    "paths": {
...
        "/reviews/book/{id}": {
            "get": {
                "description": "Find's all book's reviews with page and pagesize arguments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Get book's reviews",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID or translit name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number, must be greater or equal than 0, optional parameter",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default 1, must be greater or equal than 1",
                        "name": "pagesize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reviews_pb.ReviewModel"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.CustomError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/shared_pb.ErrorDetail"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "No reviews found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.CustomError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/shared_pb.ErrorDetail"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal error occurred",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.CustomError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/shared_pb.ErrorDetail"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "Service does not responding (maybe crush)",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.CustomError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/shared_pb.ErrorDetail"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "middleware.CustomError": {
            "description": "General error object. This structure always returns when error occurred",
            "type": "object",
            "properties": {
                "code": {
                    "description": "Internal gRPC error code (e.g. 3)",
                    "type": "integer",
                    "example": 3
                },
                "details": {
                    "description": "Error details. Check 'ErrorDetail' structure for more information",
                    "type": "array",
                    "items": {}
                },
                "message": {
                    "description": "Error message. Can be shown to users",
                    "type": "string",
                    "example": "Bad token provided"
                },
                "type": {
                    "description": "Error code in string (e.g. InvalidArgument)",
                    "type": "string",
                    "example": "InvalidArgument"
                }
            }
        },
        "reviews_pb.ReviewModel": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "creator": {
                    "$ref": "#/definitions/reviews_pb.UserModel"
                },
                "downvotes": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "text": {
                    "type": "string"
                },
                "upvotes": {
                    "type": "integer"
                },
                "userAction": {
                    "$ref": "#/definitions/reviews_pb.UserActionEnum"
                }
            }
        },
        "reviews_pb.UserActionEnum": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "UserActionEnum_noAction",
                "UserActionEnum_like",
                "UserActionEnum_dislike"
            ]
        },
        "reviews_pb.UserModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                }
            }
        },
        "shared_pb.ErrorDetail": {
            "description": "Error detail contains information about error",
            "type": "object",
            "properties": {
                "actualvalue": {
                    "description": "Actual value of field that causes the error. Note: 'password' field will be hidden",
                    "type": "string",
                    "example": "token"
                },
                "description": {
                    "description": "Error description. Only development purposes, do not show users",
                    "type": "string",
                    "example": "Field must be a jwt token"
                },
                "field": {
                    "description": "Field that error occured on",
                    "type": "string",
                    "example": "Token"
                },
                "struct": {
                    "description": "Structure that contains field",
                    "type": "string",
                    "example": "users_pb.TokenRequest"
                },
                "tag": {
                    "description": "Failed validation tag",
                    "type": "string",
                    "example": "jwt"
                },
                "tagValue": {
                    "description": "Valitation tag value",
                    "type": "string",
                    "example": "5"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Cookie"
        }
    }
}